Clean up ContentControllerBase and move document related operations to DocumentControllerBase
