--- src/Umbraco.Core/Services/OperationStatus/ContentCreatingOperationStatus.cs
@@ -1,7 +1,0 @@
-﻿namespace Umbraco.Cms.Core.Services.OperationStatus;
-
-public enum ContentCreatingOperationStatus
-{
-    Success,
-    NotFound
-}
